requirements.txt
numpy
xgboost
jax
pandas
datasets
optax
scikit-learn
matplotlib
IPython
slune-lib